pyproject.toml
[tool.poetry]
name = "hase_discord_bot"
version = "2.0.0"
description = "Basic Python based discord bot with various features."
authors = ["danczw <dev.dc@pm.me>"]
license = "Apache License 2.0"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.0"
openai = "^0.27.0"
"discord.py" = "^2.1.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
ruff = "^0.0.247"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
